+++ components/tabs/Cargo.toml
@@ -6,6 +6,9 @@ authors = ["application-services@mozilla.com"]
 license = "MPL-2.0"
 exclude = ["/android", "/ios"]
 
+[features]
+full-sync = ["sync15"]
+
 [dependencies]
 anyhow = "1.0"
 error-support = { path = "../support/error" }
@@ -18,7 +21,8 @@ serde_derive = "1"
 serde_json = "1"
 sql-support = { path = "../support/sql" }
 sync-guid = { path = "../support/guid", features = ["random"] }
-sync15 = { path = "../sync15" }
+sync15 = { path = "../sync15", optional = true }
+sync15-traits = { path = "../support/sync15-traits" }
 thiserror = "1.0"
 uniffi = "^0.19"
 uniffi_macros = "^0.19"
