+++ components/tabs/src/error.rs
@@ -4,8 +4,14 @@
 
 #[derive(Debug, thiserror::Error)]
 pub enum TabsError {
+    #[cfg(feature = "full-sync")]
     #[error("Error synchronizing: {0}")]
     SyncAdapterError(#[from] sync15::Error),
+
+    // Note we are abusing this as a kind of "mis-matched feature" error.
+    #[cfg(not(feature = "full-sync"))]
+    #[error("Sync feature is disabled: {0}")]
+    SyncAdapterError(String),
 
     #[error("Sync reset error: {0}")]
     SyncResetError(#[from] anyhow::Error),
@@ -27,6 +33,7 @@ pub enum TabsError {
 }
 
 // Adapts errors from the sync15_traits crate into sync15 errors.
+#[cfg(feature = "full-sync")]
 impl From<sync15::SyncTraitsError> for TabsError {
     fn from(e: sync15::SyncTraitsError) -> Self {
         TabsError::SyncAdapterError(e.into())
