+++ components/tabs/src/storage.rs
@@ -182,7 +182,11 @@ impl TabsStorage {
             }
         }
     }
+}
 
+// Implementations available only when syncing is enabled.
+#[cfg(feature = "full-sync")]
+impl TabsStorage {
     pub(crate) fn replace_remote_tabs(
         &mut self,
         new_remote_tabs: Vec<ClientRemoteTabs>,
