+++ components/tabs/src/sync/engine.rs
@@ -4,19 +4,43 @@
 
 use crate::storage::{ClientRemoteTabs, RemoteTab};
 use crate::sync::record::{TabsRecord, TabsRecordTab};
-use crate::sync::store::TabsStore;
+use crate::TabsStore;
 use anyhow::Result;
+use interrupt_support::NeverInterrupts;
 use std::cell::{Cell, RefCell};
 use std::collections::HashMap;
-use std::sync::Arc;
+use std::sync::{Arc, Mutex, Weak};
 use sync15::{
     clients::{self, RemoteClient},
     telemetry, CollectionRequest, DeviceType, EngineSyncAssociation, IncomingChangeset,
     OutgoingChangeset, Payload, ServerTimestamp, SyncEngine,
 };
+use sync15::{sync_multiple, MemoryCachedState, SyncEngineId};
+
 use sync_guid::Guid;
 
 const TTL_1_YEAR: u32 = 31_622_400;
+
+// Our "sync manager" will use whatever is stashed here.
+lazy_static::lazy_static! {
+    // Mutex: just taken long enough to update the inner stuff
+    pub(super) static ref STORE_FOR_MANAGER: Mutex<Weak<TabsStore>> = Mutex::new(Weak::new());
+}
+
+/// Called by the sync manager to get a sync engine via the store previously
+/// registered with the sync manager.
+pub fn get_registered_sync_engine(engine_id: &SyncEngineId) -> Option<Box<dyn SyncEngine>> {
+    let weak = STORE_FOR_MANAGER.lock().unwrap();
+    match weak.upgrade() {
+        None => None,
+        Some(store) => match engine_id {
+            SyncEngineId::Tabs => Some(Box::new(TabsEngine::new(Arc::clone(&store)))),
+            // panicing here seems reasonable - it's a static error if this
+            // it hit, not something that runtime conditions can influence.
+            _ => unreachable!("can't provide unknown engine: {}", engine_id),
+        },
+    }
+}
 
 impl RemoteTab {
     fn from_record_tab(tab: &TabsRecordTab) -> Self {
@@ -233,6 +257,73 @@ impl SyncEngine for TabsEngine {
     }
 }
 
+impl crate::TabsStore {
+    pub fn reset(self: Arc<Self>) -> crate::error::Result<()> {
+        let engine = TabsEngine::new(Arc::clone(&self));
+        engine.reset(&EngineSyncAssociation::Disconnected)?;
+        Ok(())
+    }
+
+    /// A convenience wrapper around sync_multiple.
+    pub fn sync(
+        self: Arc<Self>,
+        key_id: String,
+        access_token: String,
+        sync_key: String,
+        tokenserver_url: String,
+        local_id: String,
+    ) -> crate::error::Result<String> {
+        let mut mem_cached_state = MemoryCachedState::default();
+        let mut engine = TabsEngine::new(Arc::clone(&self));
+
+        // Since we are syncing without the sync manager, there's no
+        // command processor, therefore no clients engine, and in
+        // consequence `TabsStore::prepare_for_sync` is never called
+        // which means our `local_id` will never be set.
+        // Do it here.
+        engine.local_id = RefCell::new(local_id);
+
+        let storage_init = &sync15::Sync15StorageClientInit {
+            key_id,
+            access_token,
+            tokenserver_url: url::Url::parse(tokenserver_url.as_str())?,
+        };
+        let root_sync_key = &sync15::KeyBundle::from_ksync_base64(sync_key.as_str())?;
+
+        let mut result = sync_multiple(
+            &[&engine],
+            &mut None,
+            &mut mem_cached_state,
+            storage_init,
+            root_sync_key,
+            &NeverInterrupts,
+            None,
+        );
+
+        // for b/w compat reasons, we do some dances with the result.
+        // XXX - note that this means telemetry isn't going to be reported back
+        // to the app - we need to check with lockwise about whether they really
+        // need these failures to be reported or whether we can loosen this.
+        if let Err(e) = result.result {
+            return Err(e.into());
+        }
+        match result.engine_results.remove("tabs") {
+            None | Some(Ok(())) => Ok(serde_json::to_string(&result.telemetry)?),
+            Some(Err(e)) => Err(e.into()),
+        }
+    }
+
+    // This allows the embedding app to say "make this instance available to
+    // the sync manager". The implementation is more like "offer to sync mgr"
+    // (thereby avoiding us needing to link with the sync manager) but
+    // `register_with_sync_manager()` is logically what's happening so that's
+    // the name it gets.
+    pub fn register_with_sync_manager(self: Arc<Self>) {
+        let mut state = STORE_FOR_MANAGER.lock().unwrap();
+        *state = Arc::downgrade(&self);
+    }
+}
+
 #[cfg(test)]
 pub mod test {
     use super::*;
@@ -303,5 +394,28 @@ pub mod test {
         assert_eq!(crt.client_name, "device with no tabs");
         assert_eq!(crt.device_type, DeviceType::Unknown);
         assert_eq!(crt.remote_tabs.len(), 0);
+    }
+
+    #[test]
+    fn test_sync_manager_registration() {
+        let store = Arc::new(TabsStore::new_with_mem_path("test"));
+        assert_eq!(Arc::strong_count(&store), 1);
+        assert_eq!(Arc::weak_count(&store), 0);
+        Arc::clone(&store).register_with_sync_manager();
+        assert_eq!(Arc::strong_count(&store), 1);
+        assert_eq!(Arc::weak_count(&store), 1);
+        let registered = STORE_FOR_MANAGER
+            .lock()
+            .unwrap()
+            .upgrade()
+            .expect("should upgrade");
+        assert!(Arc::ptr_eq(&store, &registered));
+        drop(registered);
+        // should be no new references
+        assert_eq!(Arc::strong_count(&store), 1);
+        assert_eq!(Arc::weak_count(&store), 1);
+        // dropping the registered object should drop the registration.
+        drop(store);
+        assert!(STORE_FOR_MANAGER.lock().unwrap().upgrade().is_none());
     }
 }
