+++ examples/tabs-sync/Cargo.toml
@@ -21,7 +21,7 @@ path = "src/tabs-sync.rs"
 
 [dev-dependencies]
 base64 = "0.13"
-tabs = { path = "../../components/tabs" }
+tabs = { path = "../../components/tabs", features=["full-sync"] }
 serde_json = "1"
 log = "0.4"
 anyhow = "1.0"
